fix: fixing the association config

+++ configuration.py
@@ -1,5 +1,6 @@
 from camel_case import to_camel_case
 import os
+from utils import insert_lines_v2
 
 def generate_key(entity,key):
     filename = 'templates/configuration_template.txt'
@@ -17,7 +18,7 @@ def generate_key(entity,key):
         f.write(new_file_content)
         
 
-def generate_association(entity1, primary_key, entity2, entity3):
+def generate_sponsoring_association(entity1, primary_key, entity2, entity3):
     filename = 'templates/configuration_template.txt'
 
     with open(filename, 'r') as f:
@@ -46,7 +47,64 @@ def generate_association(entity1, primary_key, entity2, entity3):
 
     # Define the new file name
     new_filename = os.path.join('Configurations',f"{to_camel_case(entity1)}Configuration.cs")
+    if (os.path.isfile(new_filename)):
+        insert_lines_v2("//add config", sponsoring_config, new_filename)
+    # Write the updated file content to the new file
+    else:
+        with open(new_filename, 'w') as f:
+            f.write(new_file_content)
+
+
+def generate_many_to_one(entity1, entity2):
+    filename = 'templates/configuration_template.txt'
+
+    with open(filename, 'r') as f:
+        file_content = f.read()
+
+    sponsoring_config=f"\t\tbuilder.HasOne(p => p.{entity2})\n"
+    sponsoring_config+=f"\t\t\t.WithMany(p => p.{entity1}s)\n"
+    sponsoring_config+=f"\t\t\t.HasForeignKey(p => p.{entity2}Fk)\n"
+    sponsoring_config+=f"\t\t\t.OnDelete(DeleteBehavior.Restrict);\n\n"
+
+    
+
+    # Replace all occurrences of '{entity}' with the new entity name
+    new_file_content = file_content.replace('{entity}', entity1)
+
+    # Define the new file name
+    new_filename = os.path.join('Configurations',f"{to_camel_case(entity1)}Configuration.cs")
 
     # Write the updated file content to the new file
-    with open(new_filename, 'w') as f:
-        f.write(new_file_content)
+    if (os.path.isfile(new_filename)):
+        insert_lines_v2("//add config", sponsoring_config, new_filename)
+    # Write the updated file content to the new file
+    else:
+        with open(new_filename, 'w') as f:
+            f.write(new_file_content)
+    
+    insert_lines_v2("//add config", sponsoring_config, new_filename)
+
+def generate_many_to_many(entity1, entity2):
+    filename = 'templates/configuration_template.txt'
+
+    with open(filename, 'r') as f:
+        file_content = f.read()
+
+    sponsoring_config=f"\t\tbuilder.HasMany(p => p.{entity2}s)\n"
+    sponsoring_config+=f"\t\t\t.WithMany(p => p.{entity1}s)\n"
+    sponsoring_config+=f"\t\t\tUsingEntity(j => j.ToTable(\"{entity1}{entity2}\"))\n\n"
+
+
+    # Replace all occurrences of '{entity}' with the new entity name
+    new_file_content = file_content.replace('{entity}', entity1)
+
+    # Define the new file name
+    new_filename = os.path.join('Configurations',f"{to_camel_case(entity1)}Configuration.cs")
+
+    # Write the updated file content to the new file
+    if (os.path.isfile(new_filename)):
+        insert_lines_v2("//add config", sponsoring_config, new_filename)
+    # Write the updated file content to the new file
+    else:
+        with open(new_filename, 'w') as f:
+            f.write(new_file_content)
